KVM: PPC: Book3S PR: Fix size of XER field in shadow_vcpu
Commit 26d43146e3d9 ("KVM: PPC: Book3S: Correct width in XER
handling") changed the code to use 64-bit loads and stores (on
64-bit implementations) but forgot to change the xer field in
struct kvmppc_book3s_shadow_vcpu to 64-bit also.  This remedies
the omission.

+++ arch/powerpc/include/asm/kvm_book3s_asm.h
@@ -111,7 +111,7 @@ struct kvmppc_host_state {
 struct kvmppc_book3s_shadow_vcpu {
 	ulong gpr[14];
 	u32 cr;
-	u32 xer;
+	ulong xer;
 	ulong ctr;
 	ulong lr;
 	ulong pc;
